Updated according to suggestions, checked into SVN (.PNG attached).
"Box" is a subtype of release, with 2..* Releases. Strictly speaking, model also allows Box to have Recordings, but that's merely a simplication in the model. 

Simplified the MasterRelease pattern, Release now has an association to itself. A Release can have many other Release attached, and 0..1 Master. Without a Master, it is a MasterRelease. The same applies to a Box (MasterBox with possibly other release - even if it's not very interesting).

Brought back the Album/Single/Compilation concepts - a Release realizes these concepts.

Simplified the Subjective Model slightly (only diagram changes, suppressed some un-interesting details).