Initial stub with some SMD-related behavior. Basically a skeleton, but it does actually connect to the server, reads Artists and Releases, and presents them in a simple editor  that lets you edit the name.
There are a gazillion things missing. For instance, the SMD server code is checked in as a third party lib; it should be built or at least downloaded. The server-connection parameters are hard-coded. It does not traverse Entity relations. There is no way of saving any data.

I hope to add some more stuff as we, and in time this may or may not evolve to something that resembles a real application. But until that happens (if ever), here 's a general disclaimer applying to all commits until further notice:

*** This is a prototype ***

You have been warned.