Update to match new server version with m:n relation between  Recording and Work.
NOTE: CODE COMPILES AND RUNS BUT NO TRACKS ARE SHOWN!

Not sure why this is, the "rich" Recording does not contain any tracks when fetched from server during "inflate".

Run in debug mode and open a Release with a breakpoint set at the entry of this method:
    org.socialmusicdiscovery.rcp.content.DataSource.inflate(ObservableEntity<T>)

Inspect the "richEntity" that is returned; it holds no tracks.