RCP Client: now more or less handles "delete", but not (yet) "add" - you should be able to delete Artists, Releases, Tracks and Contributors, and all changes should be immediately reflected in the UI. For instance, if you have the Artist editor open and delete a contributor in the Release editor, the Artist contributions are instantly updated.
However, there are a few oddities and bugs. I have not sorted them all out, since I think we need to discuss how to deal with them the best way.

Here's the issue:
When you delete a Track from a Release, the track (being a first class Entity) is immediately deleted from the server and DB. The change is immediate, and the Release is not marked dirty, nor does it need to be saved to persist the change. However, deleting a Contributor from the Release makes the Release dirty, and the contributor is not deleted until the Release is saved. Besides the mere technical aspects, this is a counter-intuitive behavior from a user perspective.

There is also another bug that I just discovered; when manually connecting to the server , client will crash with the following error: 
org.eclipse.core.runtime.AssertionFailedException: assertion failed: Getter called on disposed observable org.eclipse.core.databinding.observable.set.WritableSet@151dcd6
This happens if running with "autoconnect" unchecked, or if disconnecting and reconnecting. 

Details:
- most/all classes implement Set properties as a WritableSet
- delete ObservableContribution
- ObservableContributor keeps track of entity
- ObservableContributor.delete() asks owner entity to remove contributor
- ObservableTrack.delete() asks Release and Recording to remove track
- ObservableTrack.MyContributorFacade.setContributors() reimplemented to avoid making track dirty 
- add Deletable.getDependentsToDelete()
- add Delete.resolveDependents()
- AbstractContributableEntity updates dependents (e.g. Artist.getCOngtributions())
- AbstractObservableEntity lets subclasses disable and enable "dirty" flag while setting derive properites
- AbstractObservableEntity keeps track of the DataSource
- Root implements children as a WritableSet (apparently with some bugs)
- editors adapted to use WritableSet properties directly
- add AbstractObservable.updateList(...
- JSONProvider.getConversionMap) maps WritableSet and WritableList
- create GenericWritableList and GenericWritableSet
- build.properties includes source for 3rd party jars (not sure if this has any effect)
- relocate AbstractTestCase, adapt, extend and add more unit tests